Use availability data from civTechTrees.json
This enables us to update the availability of units, techs, and buildings
with a script instead of manual edits (as long as the game developers
update civTechTrees.json correctly, that is).

To facilitate this, a few changes had to be made:
- Use units 109 and 621 for the two town centers (instead of 621 for both)
- Use unit 487 instead of 78 for Gate
- Use unit 792 instead of 790 for Palisade Gate
- Use units 1225 and 1227 for the Konniks from the Krepost
- Remove custom african and asian monk visuals (those were just a cool a WK remnant anyway)
- Use unit 1258 instead of 35 for Battering Ram
- Use unit 331 instead of 42 for Trebuchet (i.e. packed instead of unpacked)

The data file now contains a techtrees section, which lists the **available** units,
buildings, and techs for each civilisation as well as the unique units and unique techs from
the castle.

A lot of the code to find out which unit/building/tech is available to a specific civilisation
has therefore been able to be deleted. Everything should be a lot simpler now.