Remove animation from crosses
When typing in the civ selection box to select a specific civ, it could
happen that the animated fade-out of the crosses would interfere with
setting them visible again, leading to missing crosses in the end
result.

Example: Starting with the Aztecs and typing "Bo" to select Bohemians
first selects Berbers, starting an animation to remove the cross from
Thumb Ring. Then, Bohemians get selected, the cross is still opaque so
it does not reset, and when the animation finishes the cross is no
longer visible despite Bohemians not getting Thumb Ring currently.

Removing the animation from the crosses resolves this issue.