Correct wrong translation
The English text of this Japanese bonus is 
"Infantry attack 33% faster starting in Feudal Age"
An exact translation to German would be:
"Infantrie attackiert 33% schneller beginnend in der Feudalzeit"
However, the text is: 
"Angriffsgeschw. ab Feudalzeit +33 %"
As yo ucan see, it does not mention the attack speed bonus only applies to infantry units. This is not a msitake of your data but of the original game data, which simply is an incorrect translation.

I suggest to fix this here as it's simply wrong, the bonus is not for all units but only for infantry units.